feat(insights): PROBLEM type, triage fields, kanban view; rename /research→/insights
B3: surface the general triage fields (impact/confidence/category/evidence/
parked) and the PROBLEM type in the Insights list + create modal; add park/unpark
and a Show-parked toggle (includeParked).
B4: add a List/Board toggle; the board renders the four InsightStatus columns via
the shared <KanbanBoard> with a bespoke InsightCard.
B5: move the insights list from /research to /insights; /research and /problems
redirect to /insights (Research *session* subroutes untouched); drop the Problems
tab and the old ProblemKanbanBoard.

Adds insight.test.ts covering the triage fields, park/unpark, and parked filter.

Refs: Clear #28, Clear #29, Clear #30
Exponential-Ticket: cmr24pkja000pic049gpa5x6b
